Stats function moved to traixroute_output class.
