+++ lib/traixroute/application.py
@@ -39,11 +39,9 @@ import ujson
 import signal
 import threading
 import math
-import copy
 from distutils.dir_util import copy_tree
 from shutil import copyfile
 import xmlrpc.client
-import copy
 
 
 class traIXroute():
@@ -342,37 +340,7 @@ class traIXroute():
             
             # Extracting statistics.
             if enable_stats:
-                self.stats_extract(homepath, num_ips, self.detection_rules.rules, final_rules_hit, exact_time)
-
-    def stats_extract(self, homepath, num_ips, rules, final_rules_hit, time):
-        '''
-        Writes various statistics to the stats.txt file.
-        Input:
-            a) homepath: The home directory path of traIXroute.
-            b) num_ips: The number of IPs to send probes.
-            c) rules: The rules that detected IXP crossing links.
-            d) funal_rules_hit: The number of "hits" for each rule.
-            e) time: The starting timestamp of traIXroute.
-        '''
-
-        num_hits = sum(final_rules_hit)
-        
-        if num_ips:
-            with open(homepath+'/output/output_' + time + '.stats', 'w') as fp_stats:
-                temp = num_hits / num_ips
-                data = 'traIXroute stats from ' + time + ' to ' + datetime.datetime.now().strftime("%Y_%m_%d-%H_%M_%S") + '\n' +\
-                    'Number of IXP hits: ' + str(num_hits) + '\n' +\
-                    'Number of traIXroutes: ' + str(num_ips) + '\n' +\
-                    'IXP hit ratio: ' + str(temp) + '\n' + \
-                    'Number of hits per rule:\n'
-                for myi in range(0, len(rules)):
-                    if num_hits > 0:
-                        temp = final_rules_hit[myi] / num_hits
-                        data += 'Rule ' + str(myi + 1) + ': Times encountered: ' + str(
-                            final_rules_hit[myi]) + ' - Encounter Percentage: ' + str(temp) + '\n'
-                    else:
-                        data += 'Rule ' + str(myi + 1) + ': Times encountered:0 Encounter Percentage:0\n'
-                fp_stats.write(data)
+                output.stats_extract(homepath, num_ips, self.detection_rules.rules, final_rules_hit, exact_time)
             
 def run_traixroute():
     traIXroute_module = traIXroute()
